requirements.txt
amqp==2.5.1
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.1
django-celery-beat==1.5.0
django-crispy-forms==1.8.0
django-phonenumber-field==3.0.1
django-timezone-field==3.1
django-twilio==0.11.0
flower==0.9.3
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
packaging==19.2
phonenumbers==8.10.21
Pillow==6.2.1
pluggy==0.13.0
psycopg2-binary==2.8.4
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.2
pytest-django==3.6.0
python-crontab==2.4.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sorl-thumbnail==12.5.0
supervisor==4.1.0
tornado==5.1.1
twilio==6.32.0
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
zipp==0.6.0
